Updated test command line parsing to support commmon commands
