+++ test/recipes/90-test_includes.t
@@ -22,4 +22,4 @@ if ($^O eq "VMS") {
     ok(run(test(["conf_include_test", data_file("vms-includes-file.cnf")])),
        "test file includes, VMS syntax");
 }
-ok(run(test(["conf_include_test", data_file("includes-broken.cnf"), "f"])), "test broken includes");
+ok(run(test(["conf_include_test", "-f", data_file("includes-broken.cnf")])), "test broken includes");
