+++ test/v3ext.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2017 The OpenSSL Project Authors. All Rights Reserved.
+ * Copyright 2016-2018 The OpenSSL Project Authors. All Rights Reserved.
  *
  * Licensed under the Apache License 2.0 (the "License").  You may not use
  * this file except in compliance with the License.  You can obtain a copy
@@ -36,6 +36,8 @@ end:
     X509_free(x);
     return ret;
 }
+
+OPT_TEST_DECLARE_USAGE("cert.pem\n")
 
 int setup_tests(void)
 {
